tweaks: add logic to videoCardList
