+++ package.json
@@ -36,6 +36,6 @@
     "eslint-plugin-react-refresh": "^0.4.5",
     "postcss": "^8.4.32",
     "tailwindcss": "^3.3.6",
-    "vite": "^5.0.5"
+    "vite": "^5.0.6"
   }
 }
